Modification of controllers and models
The models and controllers for the website has been compiled.

MY_CONTROLLER : calling page building and what data to load for
Portfolio page to ‘Add’ and ‘View’

System.php : Setting up the redirection for login page

Portfolio_Controller : Connecting with database and adding the inputs
by the user into the targeted path (not yet ready or complete from
database side)

routes : routes to login page with register, login and logout. Setting
up a default controller that is linked to home(index page)